requirements.txt
flask
dart-tools
httpx
attrs
openapi-core
pyyaml